+++ js/cli-gate.js
@@ -1,0 +1,12 @@
+// import { createRequire } from "module";
+// const require = createRequire(import.meta.url);
+
+function AuthMCL(email, pass) {
+	return require("minecraft-launcher-core").Authenticator.getAuth(email, pass);
+}
+
+function AuthOff(user) {
+	return require("minecraft-launcher-core").Authenticator.getAuth(user);
+}
+
+module.exports = { AuthOff, AuthMCL };
